+++ app/Http/Controllers/Api/V1/CategoryController.php
@@ -4,6 +4,7 @@ namespace App\Http\Controllers\Api\V1;
 
 use App\Http\Controllers\Controller;
 use App\Http\Requests\Api\V1\Category\CategoryStoreRequest;
+use App\Http\Requests\Api\V1\Category\CategoryUpdateRequest;
 use App\Http\Resources\Api\V1\category\CategoryIndexResource;
 use App\Http\Resources\Api\V1\category\CategoryShowResource;
 use App\Models\Category;
@@ -41,9 +42,12 @@ class CategoryController extends Controller
     /**
      * Update the specified resource in storage.
      */
-    public function update(Request $request, string $id)
+    public function update(CategoryUpdateRequest $request, Category $category)
     {
-        //
+        $category->update([
+            "name" => $request->validated("name")
+        ]);
+        return response()->json(["category" => new CategoryShowResource($category)]);
     }
 
     /**
@@ -51,7 +55,11 @@ class CategoryController extends Controller
      */
     public function destroy(Category $category)
     {
-        $category->delete();
-        return response()->json(["message" => "success"], 200);
+        if($category->has("products")){
+            return response()->json(["message" => "This category has products"], 400);
+        }else{
+            $category->delete();
+            return response()->json(["message" => "success"], 200);
+        }
     }
 }
